fds_backend_beta/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
factory_boy==3.3.3
Faker==35.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==1.24.4
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.44.0
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
